Move clock to helpers package. Update some network tests.
