+++ sdk/core/src/main/java/com/klaviyo/coresdk/helpers/Clock.kt
@@ -1,4 +1,4 @@
-package com.klaviyo.coresdk.networking
+package com.klaviyo.coresdk.helpers
 
 internal interface Clock {
