+++ sdk/core/src/test/java/com/klaviyo/coresdk/helpers/StaticClock.kt
@@ -1,14 +1,10 @@
 package com.klaviyo.coresdk.helpers
 
-import com.klaviyo.coresdk.networking.Clock
-
 /**
  * Implementation of Clock for unit tests
  *
  * @property time
  */
-internal class StaticClock(
-    private val time: Long
-) : Clock {
+internal class StaticClock(private val time: Long) : Clock {
     override fun currentTimeMillis(): Long = time
 }
